% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cluster_plot}
\alias{cluster_plot}
\title{Create final scatter plot of kmeans clustering}
\usage{
cluster_plot(selected_k, km_list, pca, cell_type)
}
\arguments{
\item{selected_k}{Kmeans Value Selected}

\item{km_list}{List of K Means clustering results with various K values}

\item{pca}{PCA values dataframe}

\item{cell_type}{Cell type labels provided}
}
\value{
Resulting Cluster Plot
}
\description{
Create final scatter plot of kmeans clustering
}
\examples{
library(analysiskmeans)
data(example_sce)
SingleCellExperiment::logcounts(example_sce) <- log1p(SingleCellExperiment::counts(example_sce))#Line taken from Claude AI - Debug
sce <- example_sce
results <- data_config(sce)
mat_norm <- top_x_genes(results$sce, n_top = 50)
pca <- computepca(mat_norm)
outputs <- k_means(4, 9, n_starts = 25, seed = 42, pca)
cluster_plot(7, outputs$km_list, pca, results$cell_type)
}
