% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_config.r
\name{data_config}
\alias{data_config}
\title{Configure and Normalize input data}
\usage{
data_config(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment object}
}
\value{
A SingleCellExperiment subset to the top n_top variable genes as a matrix

Corresponding cell type labels

Log Normalized SingleCellExperiment object
}
\description{
Configure and Normalize input data
}
