% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{elbow_plot}
\alias{elbow_plot}
\title{Conduct K Means Clustering}
\usage{
elbow_plot(metrics)
}
\arguments{
\item{mat_norm}{A normalized matrix of the top x genes you want to analyze}

\item{max_k}{The max amount k value you wish to go up to}

\item{pca_mat}{Resulting PCA matrix}

\item{seed}{Random seed for reproducibility}

\item{n_starts}{Number of initial sets}
}
\value{
PCA matrix
}
\description{
Conduct K Means Clustering
}
