% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{evaluation_metrics}
\alias{evaluation_metrics}
\title{Compute ARI Values for each K}
\usage{
evaluation_metrics(min_k, max_k, km_list, cell_type)
}
\arguments{
\item{min_k}{The min amount k value you wish to start at}

\item{max_k}{The max amount k value you wish to go up to}

\item{km_list}{List of K Means clustering results with various K values}

\item{cell_type}{Corresponding cell type labels}
}
\value{
ARI values for each K
}
\description{
Compute ARI Values for each K
}
