% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_config.r
\name{export_files}
\alias{export_files}
\title{Export tables of resulting metric data to files}
\usage{
export_files(sce, cell_type, km_list, selected_k, metrics, evaluation)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{cell_type}{Corresponding cell type labels}

\item{km_list}{List of K Means clustering results with various K values}

\item{selected_k}{Selected value of K for clustering}

\item{metrics}{Silhouette scores}

\item{evaluation}{ARI values for each K}
}
\value{
None
}
\description{
Export tables of resulting metric data to files
}
