% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_config.r
\name{top_x_genes}
\alias{top_x_genes}
\title{Select n top variable genes and data configuration}
\usage{
top_x_genes(sce, n_top = 100, assay_name = "counts")
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{n_top}{Number of top variable genes to select (default: 100)}

\item{assay_name}{Name of assay to use (default: "counts")}
}
\value{
Matrix of normalized top gene values
}
\description{
Select n top variable genes and data configuration
}
