% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{cluster_plot}
\alias{cluster_plot}
\title{Create final scatter plot of kmeans clustering}
\usage{
cluster_plot(selected_k, km_list, pca, cell_type)
}
\arguments{
\item{selected_k}{Kmeans Value Selected}

\item{km_list}{List of K Means clustering results with various K values}

\item{pca}{PCA values dataframe}

\item{cell_type}{Cell type labels provided}
}
\value{
Resulting Cluster Plot
}
\description{
Create final scatter plot of kmeans clustering
}
