% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{computepca}
\alias{computepca}
\title{Compute Principal Component Analysis}
\usage{
computepca(mat_norm)
}
\arguments{
\item{mat_norm}{A normalized matrix of the top x genes you want to analyze}
}
\value{
Resulting PCA matrix
}
\description{
Compute Principal Component Analysis
}
\examples{
library(analysiskmeans)
data(example_sce)
SingleCellExperiment::logcounts(example_sce) <- log1p(SingleCellExperiment::counts(example_sce))#Line taken from Claude AI - Debug
sce <- example_sce
results <- data_config(sce)
mat_norm <- top_x_genes(results$sce, n_top = 50)
computepca(mat_norm)
}
