% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{computepca}
\alias{computepca}
\title{Compute Principal Component Analysis}
\usage{
computepca(mat_norm)
}
\arguments{
\item{mat_norm}{A normalized matrix of the top x genes you want to analyze}
}
\value{
Resulting PCA matrix
}
\description{
Compute Principal Component Analysis
}
