% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_config.r
\name{data_config}
\alias{data_config}
\title{Configure and Normalize input data}
\usage{
data_config(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment object}
}
\value{
List of A SingleCellExperiment subset, List of cell type labels and Log Normalized SingleCellExperiment object
}
\description{
Configure and Normalize input data
}
\examples{
# Assuming 'sce' is a SummarizedExperiment
library(analysiskmeans)
data(example_sce)
sce <- example_sce
analysiskmeans::data_config(sce)
}
