% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_config.r
\name{data_config}
\alias{data_config}
\title{Configure and Normalize input data}
\usage{
data_config(sce)
}
\arguments{
\item{sce}{A SingleCellExperiment object}

\item{n_top}{Number of top variable genes to select (default: 100)}

\item{assay_name}{Name of assay to use (default: "counts")}

\item{counts_mat}{Matrix of counts resulting from data_config()}
}
\value{
A SingleCellExperiment subset to the top n_top variable genes as a matrix

Corresponding cell type labels

Log Normalized SingleCellExperiment object

Select n top variable genes and data configuration

Matrix of normalized top gene values
}
\description{
Configure and Normalize input data
}
\examples{
# Assuming 'sce' is a SingleCellExperiment object
# Assuming 'counts_mat' is a count matrix resulting from data_config()

data_top <- top_x_genes(sce, counts_mat, n_top = 100)
}
