% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{elbow_plot}
\alias{elbow_plot}
\title{Create an elbow plot}
\usage{
elbow_plot(metrics)
}
\arguments{
\item{metrics}{List of silhouette scores}
}
\value{
Resulting Elbow Plot
}
\description{
Create an elbow plot
}
