% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{gap_statistic}
\alias{gap_statistic}
\title{Compute Gap Statistic for each K}
\usage{
gap_statistic(pca_mat, n_starts, max_k, metrics)
}
\arguments{
\item{pca_mat}{PCA matrix}

\item{n_starts}{Number of initial sets}

\item{max_k}{The max amount k value you wish to go up to}

\item{metrics}{metrics about}

\item{km_list}{List of K Means clustering results with various K values}

\item{cell_type}{Corresponding cell type labels}
}
\value{
Table of Gap statistic values for each K

ARI values for each K
}
\description{
Compute Gap Statistic for each K
}
