% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{gap_statistic}
\alias{gap_statistic}
\title{Compute Gap Statistic for each K}
\usage{
gap_statistic(pca, n_starts = 25, max_k, metrics)
}
\arguments{
\item{n_starts}{Number of initial sets}

\item{max_k}{The max amount k value you wish to go up to}

\item{metrics}{List of silhouette scores}

\item{pca_mat}{PCA matrix}
}
\value{
Table of Gap statistic values for each K
}
\description{
Compute Gap Statistic for each K
}
