% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeans.R
\name{k_means}
\alias{k_means}
\title{Conduct K Means Clustering}
\usage{
k_means(max_k, n_starts = 25, seed = 42, pca)
}
\arguments{
\item{max_k}{The max amount k value you wish to go up to}

\item{n_starts}{Number of initial sets}

\item{seed}{Random seed for reproducibility}

\item{pca}{Resulting PCA matrix}
}
\value{
Average silhouette scores

List of K Means clustering results with various K values
}
\description{
Conduct K Means Clustering
}
